RG-54 Multiple error subscribe / unsubscribe

+++ server/src/main/java/com/rolaface/controllers/FlexErrorSubscribeController.java
@@ -7,6 +7,7 @@ import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.rolaface.entities.FlexErrorSubscribe;
@@ -31,18 +33,25 @@ public class FlexErrorSubscribeController {
 	@PostMapping
 	public FlexErrorSubscribe create(@RequestBody int errid) {
 		int userId = ((ContextUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId();
-		FlexErrorSubscribe flexErrorSubscribe = new FlexErrorSubscribe();
-		flexErrorSubscribe.setErrid(errid);
-		flexErrorSubscribe.setUserid(userId);
-		flexErrorSubscribe.setSubscribedTimestamp(new Date());
-		return flexErrorSubscribeService.create(flexErrorSubscribe);
+		FlexErrorSubscribe flexErrorSubscribe = flexErrorSubscribeService.findSubscription(errid, userId);
+		if (flexErrorSubscribe == null) {
+			flexErrorSubscribe = new FlexErrorSubscribe();
+			flexErrorSubscribe.setErrid(errid);
+			flexErrorSubscribe.setUserid(userId);
+			flexErrorSubscribe.setSubscribedTimestamp(new Date());
+			flexErrorSubscribe = flexErrorSubscribeService.create(flexErrorSubscribe);
+		}
+		return flexErrorSubscribe;
 	}
 
 	@DeleteMapping(path = { "/{id}" })
 	public void delete(@PathVariable("id") int id) {
 		int userId = ((ContextUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId();
 		FlexErrorSubscribe flexErrorSubscribe = flexErrorSubscribeService.findSubscription(id, userId);
-		flexErrorSubscribeService.delete(flexErrorSubscribe);
+		if (flexErrorSubscribe != null) {
+			flexErrorSubscribe = flexErrorSubscribeService.findSubscription(id, userId);
+			flexErrorSubscribeService.delete(flexErrorSubscribe);
+		}
 	}
 
 	@GetMapping(value = "/subscribederrors")
@@ -54,6 +63,50 @@ public class FlexErrorSubscribeController {
 			subscribedErrors.add(String.valueOf(flexError.getErrid()));
 		}
 		return subscribedErrors;
+	}
+
+	@Transactional
+	@GetMapping(value = "/subscribeerrors", params = "errids")
+	public int subscribeErrors(@RequestParam("errids") String errids) {
+		int userId = ((ContextUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId();
+		int noOfErrorSubscribed = 0;
+		try {
+			for (String errid : errids.split(",")) {
+				FlexErrorSubscribe flexErrorSubscribe = flexErrorSubscribeService
+						.findSubscription(Integer.parseInt(errid), userId);
+				if (flexErrorSubscribe == null) {
+					flexErrorSubscribe = new FlexErrorSubscribe();
+					flexErrorSubscribe.setErrid(Integer.parseInt(errid));
+					flexErrorSubscribe.setUserid(userId);
+					flexErrorSubscribe.setSubscribedTimestamp(new Date());
+					flexErrorSubscribeService.create(flexErrorSubscribe);
+					noOfErrorSubscribed++;
+				}
+			}
+		} catch (Exception e) {
+			// TODO : ExceptionHandling
+		}
+		return noOfErrorSubscribed;
+	}
+
+	@Transactional
+	@GetMapping(value = "/unsubscribeerrors", params = "errids")
+	public int unsubscribeErrors(@RequestParam("errids") String errids) {
+		int userId = ((ContextUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId();
+		int noOfErrorUnsubscribed = 0;
+		try {
+			for (String errid : errids.split(",")) {
+				FlexErrorSubscribe flexErrorSubscribe = flexErrorSubscribeService
+						.findSubscription(Integer.parseInt(errid), userId);
+				if (flexErrorSubscribe != null) {
+					flexErrorSubscribeService.delete(flexErrorSubscribe);
+					noOfErrorUnsubscribed++;
+				}
+			}
+		} catch (Exception e) {
+			// TODO : ExceptionHandling
+		}
+		return noOfErrorUnsubscribed;
 	}
 
 }
